# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "File sharing service. Upload any file, get a shareable link.",
        "title": "Trove API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/upload": {
            "post": {
                "description": "Upload a file and get a shareable link. Supports custom slugs and overwriting existing files.",
                "tags": [
                    "files"
                ],
                "summary": "Upload a file",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "description": "File to upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "slug": {
                                        "description": "Custom URL slug (lowercase alphanumeric and hyphens, 1-64 chars)",
                                        "type": "string"
                                    },
                                    "overwrite": {
                                        "description": "Set to 'true' to replace an existing file at the same slug",
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/cmd_server.UploadResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/cmd_server.ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Slug already taken",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/cmd_server.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/cmd_server.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/{slug}/raw": {
            "get": {
                "description": "Returns the raw file content with its original content type. Supports Range requests.",
                "tags": [
                    "files"
                ],
                "summary": "Download raw file",
                "parameters": [
                    {
                        "description": "File slug",
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "HTTP Range header for partial content",
                        "name": "Range",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File content",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "206": {
                        "description": "Partial file content",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "https://localhost:8080"
        },
        {
            "url": "http://localhost:8080"
        }
    ],
    "components": {
        "schemas": {
...
            "cmd_server.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "slug already taken"
                    }
                }
            },
...
            "cmd_server.UploadResponse": {
                "type": "object",
                "properties": {
                    "slug": {
                        "type": "string",
                        "example": "my-report"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://trove.labs.staging.gteng.co/my-report"
                    }
                }
            }
        }
    }
}